backend/requirements.txt
asgiref==3.6.0
attrs==23.1.0
Babel==2.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.8
django-admin-autocomplete-filter==0.7.1
django-admin-rangefilter==0.10.0
django-autocomplete-light==3.9.4
django-bootstrap5==23.1
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-filter==23.2
django-more-admin-filters==1.6
django-nested-admin==4.0.2
django-phonenumber-field==7.1.0
django-solo==2.0.0
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pdfkit==1.0.0
phonenumbers==8.13.11
Pillow==9.5.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-monkey-business==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.29.0
sentry-sdk==1.21.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vokativ==1.2.1
xlrd==2.0.1
xlsx2html==0.4.4
XlsxWriter==3.1.7
git+https://github.com/lukasvinclav/django-admin-numeric-filter.git@c5718f59eea66f67bbaa2ee542206719dfd43df3
unidecode==1.3.8
geopy==2.4.1
pyheif==0.8.0
groq==1.0.0
